moved modal to modals folder
